fix(ui): 改用提示展示批量规则更新状态
